fix: Address high priority issues in word highlighting feature
- Remove all console.log debug statements from production code
- Fix memory leak by adding cleanup on component unmount
- Remove duplicate useWordHighlight hook (keep only the context version)
- Move and update tests to match the context implementation
